% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findBestModel.R
\name{findBestModel}
\alias{findBestModel}
\title{Find the best model according to AIC}
\usage{
findBestModel(x, y, interval)
}
\arguments{
\item{interval:}{a vector containing the lower and the upper limits to the order of the models}

\item{x:}{the indipendent variable of the model}

\item{y:}{the dipendent variable of the model, ie the observations}
}
\value{
the fitted model (as returned by lm()).
}
\description{
This function uses Akaike's an information criterion to find the best linear
model in a given range of orders. Then returns the best model.
The variables in the returned model are named x and y, so you should use
these names if yuo want to call a predict on the model.
}
\seealso{
lm()
}

