% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findMinimumFromModel.R
\name{findMinimumFromModel}
\alias{findMinimumFromModel}
\title{Minimum from a model}
\usage{
findMinimumFromModel(model, interval)
}
\arguments{
\item{model:}{model to be optimized}

\item{interval:}{a vector containing the lower and th upper limit of the interval where the minimum has to be searched}
}
\value{
A list with components minimum and objective which give the location of the minimum and the value of the function at that point.
}
\description{
This function find the minimum of the predicted value of a given model.
The model is assumed to be described by the formula y~poly(x)
}

